Update forbidden apis plugin to 3.5.1
This version replaces the `internalRuntimeForbidden` flag with a bundle named `jdk-non-portable`
The newer version also uncoverd an issue in SingleFileLdifPartition where the RandomAccessFile uses the system file encoding.
The error has been suppressed with an annotation for now. Updating the plugin will help ensure other code doesn't creep in.

NOTE: The @SupressForbidden annotation is needed at compile time, but not runtime, so it has been marked provided (as there are reports of some plugins, like the maven-shade-plugin, not excluding "optional" dependencies)